pocs/litellm/requirements.txt
litellm>=1.0.0
python-dotenv>=0.19.0
pyyaml>=6.0.0
openai>=1.0.0
anthropic>=0.5.0
google-cloud-aiplatform>=1.35.0
